Hide larger cluster of empty expanders
It is not very helpful, if you search for cards and find 20 clusters with (0) results.
Those are removed and replaced by a placeholder. To allow a bit more context, this only happens
for clusters which have 0 cards in it AND their predecessor has 0 cards in it AND its successor as well.